pyproject.toml
[project]
name = "seniority"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.24,<1.36",
    "mypy-boto3-s3>=1.35.22,<1.36",
    "grpcio-tools>=1.66.1,<1.67",
    "grpcio>=1.66.1,<1.67",
    "protobuf>=5.27.2,<5.28.0",
    "pydantic>=2.9.2,<2.10",
    "redis>=5.0.8,<5.1",
]

[project.scripts]
server = "seniority.server:main"
client = "seniority.client:main"
sample = "seniority.sample_generator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "boto3-stubs>=1.35.22,<1.36",
    "ipython>=8.27.0",
    "mypy>=1.11.2",
    "ruff>=0.6.6",
    "types-protobuf>=5.27.0.20240920,<5.28",
]
